moved `run` fn into common, moved mock into module
